signed and unsigned fixes. removed unneeded vars.

+++ lib/x509/x509.c
@@ -529,7 +529,7 @@ gnutls_x509_crt_get_signature (gnutls_x509_crt_t cert,
 
   len = bits / 8;
 
-  if (*sizeof_sig < len)
+  if (*sizeof_sig < (unsigned int)len)
     {
       *sizeof_sig = bits / 8;
       return GNUTLS_E_SHORT_MEMORY_BUFFER;
@@ -902,7 +902,7 @@ parse_general_name (ASN1_TYPE src, const char *src_name,
 		    int seq, void *name, size_t * name_size, 
 		    unsigned int* ret_type, int othername_oid)
 {
-  int len;
+  unsigned int len;
   char nptr[MAX_NAME_SIZE];
   int result;
   opaque choice_type[128];
